splitting more components and moving the non-redux parts to a new shared lib
