+++ libs/feature-movie-list-redux/src/lib/components/movie-filter/index.ts
@@ -1,1 +1,1 @@
-export { default as MovieFilterComponent } from './MovieFilter';
+export { default as MovieFilterComponent } from './MovieFilterComponent';
