--- libs/feature-movie-list-redux/src/lib/models/list.model.ts
@@ -1,6 +1,0 @@
-import { SortModel } from './SortModel';
-
-export interface ListProps {
-  componentName: string;
-  sort: SortModel;
-}
