+++ libs/shared/models/src/index.ts
@@ -7,3 +7,4 @@ export {
   FilterSettingsRaw,
   FilterSettings,
 } from './FilterSettingsModel';
+export { SortModel } from './SortModel';
