+++ libs/shared/movie-components/src/index.ts
@@ -1,3 +1,4 @@
 export { ListFilter } from './lib/filter/ListFilter';
 export { RegularFilter } from './lib/filter/RegularFilter';
 export { ExtendedFilter } from './lib/filter/ExtendedFilter';
+export { MovieList } from './lib/movie-list/MovieList';
