+++ libs/shared/movie-components/src/lib/filter/filter.component.scss
@@ -22,8 +22,8 @@
 }
 
 .filter-header-toggle-container
-> .filter-toggles
-> .filter-header-toggle-extended,
+  > .filter-toggles
+  > .filter-header-toggle-extended,
 .filter-header-toggle-container > .filter-toggles > .filter-header-toggle {
   display: inline-flex;
   align-items: center;
