+++ libs/shared/util-helpers/README.md
@@ -1,0 +1,7 @@
+# shared-util-helpers
+
+This library was generated with [Nx](https://nx.dev).
+
+## Running unit tests
+
+Run `nx test shared-util-helpers` to execute the unit tests via [Jest](https://jestjs.io).
