+++ libs/shared/util-helpers/package.json
@@ -1,0 +1,4 @@
+{
+  "name": "@giron/shared-util-helpers",
+  "version": "0.0.1"
+}
