+++ libs/shared/util-helpers/src/index.ts
@@ -1,0 +1,1 @@
+export { scrollToTop } from './page.utils';
