+++ libs/shared/util-helpers/src/page.utils.ts
@@ -1,0 +1,3 @@
+export const scrollToTop = () => {
+  window.scrollTo(0, 0);
+};
